restore ExecuteExtensions
