+++ monix-execution/shared/src/main/scala/monix/execution/Scheduler.scala
@@ -20,7 +20,7 @@ package monix.execution
 import java.util.concurrent.Executor
 
 import monix.execution.internal.RunnableAction
-import monix.execution.schedulers.{SchedulerCompanionImpl, StartAsyncBatchRunnable, TrampolinedRunnable}
+import monix.execution.schedulers.SchedulerCompanionImpl
 
 import scala.annotation.implicitNotFound
 import scala.concurrent.ExecutionContext
@@ -304,64 +304,7 @@ object Scheduler extends SchedulerCompanionImpl {
   val TRACING = Features.flag(2)
 
   /** Utilities complementing the `Scheduler` interface. */
-  implicit final class Extensions(val source: Scheduler) extends AnyVal {
-
-    /** Schedules the given callback for asynchronous
-      * execution in the thread-pool.
-      *
-      * On Scala < 2.12 it is described as a macro, so it
-      * has zero overhead, being perfectly equivalent with
-      * `execute(new Runnable { ... })`.
-      *
-      * On Scala 2.12 because of the Java 8 SAM types integration,
-      * this extension macro is replaced with a method that takes
-      * a plain `Runnable` as parameter.
-      *
-      * @param cb the callback to execute asynchronously
-      */
-    def executeAsync(cb: Runnable): Unit =
-      source.execute(cb)
-
-    /** Schedules the given callback for asynchronous
-      * execution in the thread-pool, but also indicates the
-      * start of a
-      * [[monix.execution.schedulers.TrampolinedRunnable thread-local trampoline]]
-      * in case the scheduler is a
-      * [[monix.execution.schedulers.BatchingScheduler BatchingScheduler]].
-      *
-      * This utility is provided as an optimization. If you don't understand
-      * what this does, then don't worry about it.
-      *
-      * On Scala < 2.12 it is described as a macro, so it
-      * has zero overhead. On Scala 2.12 because of the Java 8 SAM
-      * types integration, this extension macro is replaced with a
-      * method that takes a plain `TrampolinedRunnable` as parameter.
-      *
-      * @param cb the callback to execute asynchronously
-      */
-    def executeAsyncBatch(cb: TrampolinedRunnable): Unit = {
-      val r = StartAsyncBatchRunnable(cb, source)
-      source.execute(r)
-    }
-
-    /** Schedules the given callback for immediate execution as a
-      * [[monix.execution.schedulers.TrampolinedRunnable TrampolinedRunnable]].
-      * Depending on the execution context, it might
-      * get executed on the current thread by using an internal
-      * trampoline, so it is still safe from stack-overflow exceptions.
-      *
-      * On Scala < 2.12 it is described as a macro, so it
-      * has zero overhead, being perfectly equivalent with
-      * `execute(new TrampolinedRunnable { ... })`.
-      *
-      * On Scala 2.12 because of the Java 8 SAM types integration,
-      * this extension macro is replaced with a method that takes
-      * a plain `TrampolinedRunnable` as parameter.
-      *
-      * @param cb the callback to execute asynchronously
-      */
-    def executeTrampolined(cb: TrampolinedRunnable): Unit =
-      source.execute(cb)
+  implicit final class Extensions(val source: Scheduler) extends AnyVal with schedulers.ExecuteExtensions {
 
     /** Schedules a task to run in the future, after `initialDelay`.
       *
